refactor(zk-explorer): Organize server-side functions in a single place
