+++ apps/zk-explorer/components/ProofGenerationStepCards/index.tsx
@@ -13,9 +13,9 @@ import {
 } from '@webb-tools/webb-ui-components';
 import assert from 'assert';
 import { FC, useCallback, useMemo, useState } from 'react';
-import { requestProofGeneration } from '../../api/services';
 import { CONTACT_URL } from '../../constants';
 import { MOCK_MPC_PARTICIPANTS } from '../../constants/mock';
+import { requestProofGeneration } from '../../server/services';
 import { RelativePageUrl } from '../../utils';
 import FileUploadAreaWithList from '../FileUploadAreaWithList';
 import IdentityItem from './IdentityItem';
