+++ apps/zk-explorer/containers/ProjectDetailTabs/Circuits/Circuits.tsx
@@ -1,0 +1,8 @@
+import { fetchProjectFileTree } from '../../../server/projectDetails';
+import CircuitsClient from './client';
+
+export default async function Circuits() {
+  const circuitFilesData = await fetchProjectFileTree();
+
+  return <CircuitsClient fileTree={circuitFilesData} />;
+}
