--- apps/zk-explorer/containers/ProjectDetailTabsContainer/Circuits/Circuits.tsx
@@ -1,9 +1,0 @@
-import CircuitsClient from './client';
-
-import { getProjectCircuitsData } from '../../../server';
-
-export default async function Circuits() {
-  const circuitFilesData = await getProjectCircuitsData();
-
-  return <CircuitsClient data={circuitFilesData} />;
-}
