--- apps/zk-explorer/containers/ProjectDetailTabsContainer/Circuits/index.ts
@@ -1,3 +1,0 @@
-import { default as Circuits } from './Circuits';
-
-export default Circuits;
