--- apps/zk-explorer/containers/ProjectDetailTabsContainer/Summary/index.ts
@@ -1,3 +1,0 @@
-import { default as Summary } from './Summary';
-
-export default Summary;
