--- apps/zk-explorer/containers/ProjectDetailTabsContainer/TrustedSetup/index.ts
@@ -1,3 +1,0 @@
-import { default as TrustedSetup } from './TrustedSetup';
-
-export default TrustedSetup;
