--- apps/zk-explorer/containers/ProjectDetailTabsContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectDetailTabsContainer } from './ProjectDetailTabsContainer';
