--- apps/zk-explorer/containers/ProjectGitHubInfoContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectGitHubInfoContainer } from './ProjectGitHubInfoContainer';
