--- apps/zk-explorer/containers/ProjectHeaderContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as ProjectHeaderContainer } from './ProjectHeaderContainer';
