--- apps/zk-explorer/containers/index.ts
@@ -1,4 +1,0 @@
-export { ProjectDetailTabsContainer } from './ProjectDetailTabsContainer';
-export { ProjectGitHubInfoContainer } from './ProjectGitHubInfoContainer';
-export { ProjectHeaderContainer } from './ProjectHeaderContainer';
-export { RelatedProjectsContainer } from './RelatedProjectsContainer';
