--- apps/zk-explorer/server/fetchRelatedProjectsContainerData.ts
@@ -1,23 +1,0 @@
-import { ProjectItem } from '../components/ProjectCard';
-import { DEBUG_ARTIFICIAL_DELAY_MS } from '../constants';
-import { artificialWait } from '../utils';
-
-export default async function fetchRelatedProjectsContainerData(): Promise<
-  ProjectItem[]
-> {
-  await artificialWait(DEBUG_ARTIFICIAL_DELAY_MS);
-
-  return Array(5).fill({
-    ownerAvatarUrl:
-      'https://avatars.githubusercontent.com/u/76852793?s=200&v=4',
-    repositoryOwner: 'webb',
-    repositoryName: 'masp',
-    stargazerCount: 123,
-    circuitCount: 24,
-    description:
-      'Short blurb about what the purpose of this circuit. This is a longer line to test multiline.',
-    contributorAvatarUrls: [
-      'https://avatars.githubusercontent.com/u/76852793?s=200&v=4',
-    ],
-  }) satisfies ProjectItem[];
-}
